Code restructuring. Fix csv_generator bugs.
